requirements.txt
asgiref==3.4.1
attrs==21.2.0
backports.entry-points-selectable==1.1.0
certifi==2021.5.30
cfgv==3.3.0
charset-normalizer==2.0.3
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.2
dj-database-url==0.5.0
Django==3.2.5
django-configurations==2.2
djangorestframework==3.12.4
drf-yasg==1.20.0
filelock==3.0.12
identify==2.2.13
idna==3.2
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
nodeenv==1.6.0
packaging==21.0
platformdirs==2.2.0
pluggy==0.13.1
pre-commit==2.14.0
psycopg2==2.9.1
psycopg2-binary==2.9.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
pytest-django==4.4.0
python-dotenv==0.19.0
python-json-logger==2.0.2
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.6
virtualenv==20.7.0
